refactor project view
