+++ src/main/kotlin/di/repositoryModule.kt
@@ -2,8 +2,8 @@ package org.damascus.di
 
 import logic.model.Project
 import logic.repo.DataSource
+import logic.repo.ProjectRepository
 import org.damascus.data.repo.ProjectRepositoryImpl
-import org.damascus.logic.repository.ProjectRepository
 import org.koin.dsl.module
 
 val repositoryModule = module {
