--- src/main/kotlin/ui/retrieve/PlanRetrieve.kt
@@ -1,7 +1,0 @@
-package org.damascus.ui.retrieve
-
-interface PlanRetrieve {
-    fun displayProjects()
-    fun createProject()
-    fun manageProject()
-}
