+++ src/main/kotlin/ui/views/project/ProjectView.kt
@@ -3,13 +3,12 @@ package org.damascus.ui.views.project
 import logic.model.Project
 
 interface ProjectView {
-
-    fun displayAllProjects(
+    fun displayDetailsProjects(
         projects: List<Project>,
         label: String,
         contentSelector: ((Project) -> Map<String, Any?>)? = null
     )
 
-    fun displayProjectsAsTable(projects: List<Project>)
-
+    fun displayProjects()
+    fun createProject()
 }
