+++ src/main/kotlin/ui/views/project/ProjectViewCli.kt
@@ -1,12 +1,60 @@
 package org.damascus.ui.views.project
 
+import kotlinx.datetime.Clock
+import kotlinx.datetime.TimeZone
+import kotlinx.datetime.toLocalDateTime
 import logic.model.Project
+import org.damascus.logic.usecase.ProjectUseCase.CreateProjectUseCase
+import org.damascus.logic.usecase.ProjectUseCase.GetAllProjectsUseCase
+import org.damascus.ui.io.ConsoleUserInput
 import org.damascus.ui.util.TerminalColor
 import org.damascus.ui.util.printTable
 import org.damascus.ui.util.withStyle
+import java.util.*
 
-class ProjectViewCli(): ProjectView {
-    override fun displayAllProjects(
+class ProjectViewCli(
+    private val consoleUserInput: ConsoleUserInput,
+    private val createProjectUseCase: CreateProjectUseCase,
+    private val getAllProjectsUseCase: GetAllProjectsUseCase
+) : ProjectView {
+
+    override fun displayProjects() {
+        val projects = getAllProjectsUseCase()
+        if (projects.isEmpty()) {
+            printMessageBox("No projects available.", TerminalColor.Red)
+            return
+        }
+
+        displayDetailsProjects(
+            projects = projects,
+            label = "Basic Info",
+            contentSelector = { project ->
+                mapOf(
+                    "Name" to project.name,
+                    "Mates Count" to project.assignedMatesIds.size,
+                    "Created Date" to project.creationDate
+                )
+            }
+        )
+    }
+
+    override fun createProject() {
+        val name = consoleUserInput.readString("Enter project name:")
+        val project = Project(
+            id = UUID.randomUUID(),
+            name = name,
+            assignedMatesIds = mutableListOf(),
+            creationDate = Clock.System.now().toLocalDateTime(TimeZone.UTC)
+        )
+
+        if (createProjectUseCase(project)) {
+            printMessageBox("Added Project ${project.name}")
+        } else {
+            printMessageBox("Project already exists.", TerminalColor.Red)
+        }
+    }
+
+    override fun displayDetailsProjects(
         projects: List<Project>,
         label: String,
         contentSelector: ((Project) -> Map<String, Any?>)?
@@ -21,12 +69,7 @@ class ProjectViewCli(): ProjectView {
         val headers = if (contentSelector != null) {
             listOf("ID") + contentSelector(projects.first()).keys.toList()
         } else {
-            listOf(
-                "ID",
-                "Name",
-                "Assigned Mates",
-                "Created Date"
-            )
+            listOf("ID", "Name", "Assigned Mates", "Created Date")
         }
 
         val data = projects.map { project ->
@@ -45,25 +88,6 @@ class ProjectViewCli(): ProjectView {
 
         printTable(headers, data)
     }
-
-    override fun displayProjectsAsTable(projects: List<Project>) {
-        if (projects.isEmpty()) {
-            println("❌ No projects available.")
-            return
-        }
-
-        val headers = listOf("No", "Project Name", "ID")
-        val rows = projects.mapIndexed { index, project ->
-            listOf(
-                (index + 1).toString(),
-                project.name,
-                project.id.toString()
-            )
-        }
-
-        printTable(headers, rows)
-    }
-
 
     fun printMessageBox(
         message: String,
